Add Metrics Function
Adds the Metric Function to MHDpy module.

Key trick here is the use of [:,None,None] type of reshape commands to
ensure the broadcast of unequal dim lengths works for index arrays.

Could have done this with pythonic sliceing, but getting the index
arrays working makes the rest of the code easier to import.