Add Boundary routines
Implements boundary condition routines.

Note the clever use of jc_act[:,NO:,:] to deal with the shape mismatch
issues.

At this point the change the BC type involves uncommenting code.  This
needs to be changed to a option for the function.  I believe the
currently commented out types will work, but I haven’t tested them.